test(settings): add tests
