+++ packages/settings/test/settings-spec.ts
@@ -1,0 +1,13 @@
+context("settings utilities", () => {
+  describe("getSettingsForDocument", () => {});
+
+  describe("hasSettingsForDocument", () => {});
+
+  describe("setSettingsForDocument", () => {});
+
+  describe("clearSettings", () => {});
+
+  describe("clearDocumentSettings", () => {});
+
+  describe("setGlobalSettings", () => {});
+});
